Refactor simulation.py and move all utility classes into a separte package
