+++ src/gsy_e/gsy_e_core/simulation.py
@@ -16,35 +16,30 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 
-import datetime
 import gc
 import os
-import sys
-from dataclasses import dataclass
-from importlib import import_module
 from logging import getLogger
-from time import mktime, sleep, time
-from types import ModuleType
-from typing import TYPE_CHECKING, Optional, Tuple, Union
+from time import sleep, time
+from typing import TYPE_CHECKING
 
 import psutil
 from gsy_framework.constants_limits import ConstSettings, GlobalConfig
 from gsy_framework.enums import CoefficientAlgorithm, SpotMarketTypeEnum
-from gsy_framework.kafka_communication.kafka_producer import kafka_connection_factory
 from gsy_framework.utils import format_datetime, str_to_pendulum_datetime
-from numpy import random
-from pendulum import DateTime, Duration, duration, now
+from pendulum import DateTime, Duration, duration
 
 import gsy_e.constants
-from gsy_e.constants import DATE_TIME_FORMAT, SIMULATION_PAUSE_TIMEOUT, TIME_ZONE
 from gsy_e.gsy_e_core.exceptions import SimulationException
-from gsy_e.gsy_e_core.export import CoefficientExportAndPlot, ExportAndPlot
 from gsy_e.gsy_e_core.global_objects_singleton import global_objects
-from gsy_e.gsy_e_core.live_events import LiveEvents
 from gsy_e.gsy_e_core.myco_singleton import bid_offer_matcher
-from gsy_e.gsy_e_core.redis_connections.simulation import RedisSimulationCommunication
-from gsy_e.gsy_e_core.sim_results.endpoint_buffer import (
-    CoefficientEndpointBuffer, SimulationEndpointBuffer)
+from gsy_e.gsy_e_core.simulation_utils.simulation_external_events import SimulationExternalEvents
+from gsy_e.gsy_e_core.simulation_utils.simulation_progress_info import SimulationProgressInfo
+from gsy_e.gsy_e_core.simulation_utils.simulation_results_manager import (
+    simulation_results_manager_factory, CoefficientSimulationResultsManager)
+from gsy_e.gsy_e_core.simulation_utils.simulation_setup import SimulationSetup
+from gsy_e.gsy_e_core.simulation_utils.simulation_status_manager import SimulationStatusManager
+from gsy_e.gsy_e_core.simulation_utils.simulation_time_manager import (
+    simulation_time_manager_factory)
 from gsy_e.gsy_e_core.util import NonBlockingConsole
 from gsy_e.models.area.event_deserializer import deserialize_events_to_areas
 from gsy_e.models.area.scm_manager import SCMManager
@@ -56,536 +51,8 @@ if TYPE_CHECKING:
 log = getLogger(__name__)
 
 
-RANDOM_SEED_MAX_VALUE = 1000000
-
-
 class SimulationResetException(Exception):
     """Exception for errors when resetting the simulation."""
-
-
-class SimulationProgressInfo:
-    """Information about the simulation progress."""
-
-    def __init__(self):
-        self.eta = duration(seconds=0)
-        self.elapsed_time = duration(seconds=0)
-        self.percentage_completed = 0
-        self.next_slot_str = ""
-        self.current_slot_str = ""
-        self.current_slot_time = None
-        self.current_slot_number = 0
-
-    @classmethod
-    def _get_market_slot_time_str(cls, slot_number: int, config: "SimulationConfig") -> str:
-        """Get market slot time string."""
-        return format_datetime(cls._get_market_slot_time(slot_number, config))
-
-    @staticmethod
-    def _get_market_slot_time(slot_number: int, config: "SimulationConfig") -> DateTime:
-        return config.start_date.add(
-            minutes=config.slot_length.total_minutes() * slot_number
-        )
-
-    def update(self, slot_no: int, slot_count: int, time_params: "SimulationTimeManager",
-               config: SimulationConfig) -> None:
-        """Update progress info according to the simulation progress."""
-        run_duration = (
-                now(tz=TIME_ZONE) - time_params.start_time -
-                duration(seconds=time_params.paused_time)
-        )
-
-        if gsy_e.constants.RUN_IN_REALTIME:
-            self.eta = None
-            self.percentage_completed = 0.0
-        else:
-            self.eta = (run_duration / (slot_no + 1) * slot_count) - run_duration
-            self.percentage_completed = (slot_no + 1) / slot_count * 100
-
-        self.elapsed_time = run_duration
-        self.current_slot_str = self._get_market_slot_time_str(slot_no, config)
-        self.current_slot_time = self._get_market_slot_time(slot_no, config)
-        self.next_slot_str = self._get_market_slot_time_str(
-            slot_no + 1, config)
-        self.current_slot_number = slot_no
-
-        log.warning("Slot %s of %s - (%.1f %%) %s elapsed, ETA: %s", slot_no+1, slot_count,
-                    self.percentage_completed, self.elapsed_time,
-                    self.eta)
-
-    def log_simulation_finished(self, paused_duration: Duration, config: SimulationConfig) -> None:
-        """Log that the simulation has finished."""
-        log.info(
-            "Run finished in %s%s / %.2fx real time",
-            self.elapsed_time,
-            f" ({paused_duration} paused)" if paused_duration else "",
-            config.sim_duration / (self.elapsed_time - paused_duration)
-        )
-
-
-@dataclass
-class SimulationStatusManager:
-    """State of the Simulation class."""
-    paused: bool = False
-    pause_after: duration = None
-    timed_out: bool = False
-    stopped: bool = False
-    sim_status = "initializing"
-    incremental: bool = False
-
-    @property
-    def status(self) -> str:
-        """Return status of simulation."""
-        if self.timed_out:
-            return "timed-out"
-        if self.stopped:
-            return "stopped"
-        if self.paused:
-            return "paused"
-        return self.sim_status
-
-    def stop(self) -> None:
-        """Stop simulation."""
-        self.stopped = True
-
-    @property
-    def finished(self) -> bool:
-        """Return if simulation has finished."""
-        return self.sim_status == "finished"
-
-    def toggle_pause(self) -> bool:
-        """Pause or resume simulation."""
-        if self.finished:
-            return False
-        self.paused = not self.paused
-        return True
-
-    def handle_pause_after(self, time_since_start: Duration) -> None:
-        """Deals with pause-after parameter, which pauses the simulation after some time."""
-        if self.pause_after and time_since_start >= self.pause_after:
-            self.paused = True
-            self.pause_after = None
-
-    def handle_pause_timeout(self, tick_time_counter: float) -> None:
-        """
-        Deals with the case that the pause time exceeds the simulation timeout, in which case the
-        simulation should be stopped.
-        """
-        if time() - tick_time_counter > SIMULATION_PAUSE_TIMEOUT:
-            self.timed_out = True
-            self.stopped = True
-            self.paused = False
-        if self.stopped:
-            self.paused = False
-
-    def handle_incremental_mode(self) -> None:
-        """
-        Handle incremental paused mode, where simulation is paused after each market slot.
-        """
-        if self.incremental:
-            self.paused = True
-
-
-@dataclass
-class SimulationTimeManager:
-    """Handles simulation time management."""
-    start_time: DateTime = now(tz=TIME_ZONE)
-    tick_time_counter: float = time()
-    slot_length_realtime: duration = None
-    tick_length_realtime_s: int = None
-    paused_time: int = 0  # Time spent in paused state, in seconds
-
-    def reset(self, not_restored_from_state: bool = True) -> None:
-        """
-        Restore time-related parameters of the simulation to their default values.
-        Mainly useful when resetting the simulation.
-        """
-        self.tick_time_counter = time()
-        if not_restored_from_state:
-            self.start_time = now(tz=TIME_ZONE)
-            self.paused_time = 0
-
-    def _set_area_current_tick(self, area: "Area", current_tick: int) -> None:
-        area.current_tick = current_tick
-        for child in area.children:
-            self._set_area_current_tick(child, current_tick)
-
-    def calculate_total_initial_ticks_slots(
-            self, config: SimulationConfig, slot_resume: int, tick_resume: int, area: "AreaBase"
-    ) -> Tuple[int, int, int]:
-        """Calculate the initial slot and tick of the simulation, and the total slot count."""
-        slot_count = int(config.sim_duration / config.slot_length)
-
-        if gsy_e.constants.RUN_IN_REALTIME:
-            slot_count = sys.maxsize
-
-            today = datetime.date.today()
-            seconds_since_midnight = time() - mktime(today.timetuple())
-            slot_resume = int(seconds_since_midnight // config.slot_length.seconds) + 1
-            seconds_elapsed_in_slot = seconds_since_midnight % config.slot_length.seconds
-            ticks_elapsed_in_slot = seconds_elapsed_in_slot // config.tick_length.seconds
-            tick_resume = int(ticks_elapsed_in_slot) + 1
-
-            seconds_elapsed_in_tick = seconds_elapsed_in_slot % config.tick_length.seconds
-
-            seconds_until_next_tick = config.tick_length.seconds - seconds_elapsed_in_tick
-
-            ticks_since_midnight = int(seconds_since_midnight // config.tick_length.seconds) + 1
-            self._set_area_current_tick(area, ticks_since_midnight)
-
-            sleep(seconds_until_next_tick)
-
-        if self.slot_length_realtime:
-            self.tick_length_realtime_s = (
-                    self.slot_length_realtime.seconds /
-                    config.ticks_per_slot)
-        return slot_count, slot_resume, tick_resume
-
-    def handle_slowdown_and_realtime(self, tick_no: int, config: SimulationConfig) -> None:
-        """
-        Handle simulation slowdown and simulation realtime mode, and sleep the simulation
-        accordingly.
-        """
-        if gsy_e.constants.RUN_IN_REALTIME:
-            tick_runtime_s = time() - self.tick_time_counter
-            sleep_time_s = config.tick_length.seconds - tick_runtime_s
-        elif self.slot_length_realtime:
-            current_expected_tick_time = self.tick_time_counter + self.tick_length_realtime_s
-            sleep_time_s = current_expected_tick_time - now().timestamp()
-        else:
-            return
-
-        if sleep_time_s > 0:
-            sleep(sleep_time_s)
-            log.debug("Tick %s/%s: Sleep time of %s s was applied",
-                      tick_no + 1, config.ticks_per_slot, sleep_time_s)
-
-        self.tick_time_counter = time()
-
-
-@dataclass
-class SimulationTimeManagerScm:
-    """Handles simulation time management."""
-    start_time: DateTime = now(tz=TIME_ZONE)
-    paused_time: int = 0  # Time spent in paused state, in seconds
-    slot_length_realtime: duration = None
-    slot_time_counter: float = time()
-
-    def reset(self, not_restored_from_state: bool = True) -> None:
-        """
-        Restore time-related parameters of the simulation to their default values.
-        Mainly useful when resetting the simulation.
-        """
-        self.slot_time_counter = time()
-        if not_restored_from_state:
-            self.start_time = now(tz=TIME_ZONE)
-            self.paused_time = 0
-
-    def handle_slowdown_and_realtime_scm(self, slot_no: int, slot_count: int,
-                                         config: SimulationConfig) -> None:
-        """
-        Handle simulation slowdown and simulation realtime mode, and sleep the simulation
-        accordingly for SCM simulations.
-        """
-
-        slot_length_realtime_s = self.slot_length_realtime.total_seconds()
-
-        if gsy_e.constants.RUN_IN_REALTIME:
-            slot_runtime_s = time() - self.slot_time_counter
-            sleep_time_s = config.slot_length.total_seconds() - slot_runtime_s
-        elif slot_length_realtime_s:
-            current_expected_tick_time = self.slot_time_counter + slot_length_realtime_s
-            sleep_time_s = current_expected_tick_time - now().timestamp()
-        else:
-            return
-
-        if sleep_time_s > 0:
-            sleep(sleep_time_s)
-            log.debug("Slot %s/%s: Sleep time of %s s was applied",
-                      slot_no, slot_count, sleep_time_s)
-
-        self.slot_time_counter = time()
-
-    @staticmethod
-    def calc_resume_slot_and_count_realtime(
-            config: SimulationConfig, slot_resume: int) -> Tuple[int, int]:
-        """Calculate total slot count and the slot where to resume the realtime simulation."""
-        slot_count = int(config.sim_duration / config.slot_length)
-
-        if gsy_e.constants.RUN_IN_REALTIME:
-            slot_count = sys.maxsize
-
-            today = datetime.date.today()
-            seconds_since_midnight = time() - mktime(today.timetuple())
-            slot_resume = int(seconds_since_midnight // config.slot_length.seconds) + 1
-            seconds_elapsed_in_slot = seconds_since_midnight % config.slot_length.seconds
-            sleep_time_s = config.slot_length.total_seconds() - seconds_elapsed_in_slot
-            sleep(sleep_time_s)
-
-        return slot_count, slot_resume
-
-
-def simulation_time_manager_factory(slot_length_realtime: duration):
-    """Factory for time manager objects."""
-    if ConstSettings.MASettings.MARKET_TYPE == SpotMarketTypeEnum.COEFFICIENTS.value:
-        return SimulationTimeManagerScm(slot_length_realtime=slot_length_realtime)
-    return SimulationTimeManager(slot_length_realtime=slot_length_realtime)
-
-
-@dataclass
-class SimulationSetup:
-    """Static simulation configuration."""
-    seed: int = 0
-    enable_bc: bool = False
-    use_repl: bool = False
-    setup_module_name: str = ""
-    started_from_cli: str = True
-    config: SimulationConfig = None
-
-    def __post_init__(self) -> None:
-        self._set_random_seed(self.seed)
-
-    def load_setup_module(self) -> Union["Area", "CoefficientArea"]:
-        """Load setup module and create areas that are described on the setup."""
-        loaded_python_module = self._import_setup_module(self.setup_module_name)
-        area = loaded_python_module.get_setup(self.config)
-        self._log_traversal_length(area)
-        return area
-
-    def _set_random_seed(self, seed: Optional[int]) -> None:
-        if seed is not None:
-            random.seed(int(seed))
-        else:
-            random_seed = random.randint(0, RANDOM_SEED_MAX_VALUE)
-            random.seed(random_seed)
-            seed = random_seed
-            log.info("Random seed: %s", random_seed)
-        self.seed = seed
-
-    def _log_traversal_length(self, area: "Area") -> None:
-        no_of_levels = self._get_setup_levels(area) + 1
-        num_ticks_to_propagate = no_of_levels * 2
-        time_to_propagate_minutes = (num_ticks_to_propagate *
-                                     self.config.tick_length.seconds / 60.)
-        log.info("Setup has %s levels, offers/bids need at least %s minutes to propagate.",
-                 no_of_levels, time_to_propagate_minutes)
-
-    def _get_setup_levels(self, area: "Area", level_count: int = 0) -> int:
-        level_count += 1
-        count_list = [self._get_setup_levels(child, level_count)
-                      for child in area.children if child.children]
-        return max(count_list) if len(count_list) > 0 else level_count
-
-    @staticmethod
-    def _import_setup_module(setup_module_name: str) -> ModuleType:
-        try:
-            if ConstSettings.GeneralSettings.SETUP_FILE_PATH is None:
-                return import_module(f".{setup_module_name}", "gsy_e.setup")
-            sys.path.append(ConstSettings.GeneralSettings.SETUP_FILE_PATH)
-            return import_module(f"{setup_module_name}")
-        except (ModuleNotFoundError, ImportError) as ex:
-            raise SimulationException(
-                f"Invalid setup module '{setup_module_name}'") from ex
-        finally:
-            log.debug("Using setup module '%s'", setup_module_name)
-
-
-class SimulationResultsManager:
-    """Maintain and populate the simulation results and the publishing to the message broker."""
-    def __init__(self, export_results_on_finish: bool, export_path: str,
-                 export_subdir: Optional[str], started_from_cli: bool) -> None:
-        self.export_results_on_finish = export_results_on_finish
-        self.export_path = export_path
-        self.started_from_cli = started_from_cli
-        self.kafka_connection = kafka_connection_factory()
-
-        if export_subdir is None:
-            self.export_subdir = now(tz=TIME_ZONE).format(f"{DATE_TIME_FORMAT}:ss")
-        else:
-            self.export_subdir = export_subdir
-        self._endpoint_buffer = None
-        self._export = None
-
-    def init_results(self, redis_job_id: str, area: "AreaBase",
-                     config_params: SimulationSetup) -> None:
-        """Construct objects that contain the simulation results for the broker and CSV output."""
-        self._endpoint_buffer = SimulationEndpointBuffer(
-            redis_job_id, config_params.seed,
-            area, self.export_results_on_finish)
-
-        if self.export_results_on_finish:
-            self._export = ExportAndPlot(area, self.export_path,
-                                         self.export_subdir,
-                                         self._endpoint_buffer)
-
-    @property
-    def _should_send_results_to_broker(self) -> None:
-        """Flag that decides whether to send results to the gsy-web"""
-        return not self.started_from_cli and self.kafka_connection.is_enabled()
-
-    def update_and_send_results(self, current_state: dict, progress_info: SimulationProgressInfo,
-                                area: "Area", simulation_status: str) -> None:
-        """
-        Update the simulation results. Should be called on init, finish and every market cycle.
-        """
-        assert self._endpoint_buffer is not None
-
-        if self._should_send_results_to_broker:
-            self._endpoint_buffer.update_stats(
-                area, simulation_status, progress_info, current_state,
-                calculate_results=False)
-            results = self._endpoint_buffer.prepare_results_for_publish()
-            if results is None:
-                return
-            self.kafka_connection.publish(results, current_state["simulation_id"])
-
-        elif (gsy_e.constants.RETAIN_PAST_MARKET_STRATEGIES_STATE or
-                self.export_results_on_finish):
-
-            self._endpoint_buffer.update_stats(
-                area, current_state["sim_status"], progress_info, current_state,
-                calculate_results=True)
-            self._update_area_stats(area, self._endpoint_buffer)
-
-            if self.export_results_on_finish:
-                assert self._export is not None
-                if (area.current_market is not None
-                        and gsy_e.constants.RETAIN_PAST_MARKET_STRATEGIES_STATE):
-                    # for integration tests:
-                    self._export.raw_data_to_json(
-                        area.current_market.time_slot_str,
-                        self._endpoint_buffer.flattened_area_core_stats_dict
-                    )
-
-                self._export.file_stats_endpoint(area)
-
-    @classmethod
-    def _update_area_stats(cls, area: "Area", endpoint_buffer: "SimulationEndpointBuffer") -> None:
-        for child in area.children:
-            cls._update_area_stats(child, endpoint_buffer)
-        bills = endpoint_buffer.results_handler.all_ui_results["bills"].get(area.uuid, {})
-        area.stats.update_aggregated_stats({"bills": bills})
-        area.stats.kpi.update(
-            endpoint_buffer.results_handler.all_ui_results["kpi"].get(area.uuid, {}))
-
-    def update_csv_on_market_cycle(self, slot_no: int, area: "Area") -> None:
-        """Update the csv results on market cycle."""
-        if self.export_results_on_finish:
-            self._export.data_to_csv(area, slot_no == 0)
-
-    def save_csv_results(self, area: "Area") -> None:
-        """Update the CSV results on finish, and write the CSV files."""
-        if self.export_results_on_finish:
-            log.info("Exporting simulation data.")
-            self._export.data_to_csv(area, False)
-            self._export.area_tree_summary_to_json(self._endpoint_buffer.area_result_dict)
-            self._export.export(power_flow=None)
-
-
-class CoefficientSimulationResultsManager(SimulationResultsManager):
-    """
-    Maintain and populate the simulation results and the publishing to the message broker for
-    SCM simulations.
-    """
-
-    def init_results(self, redis_job_id: str, area: "AreaBase",
-                     config_params: SimulationSetup) -> None:
-        """Construct objects that contain the simulation results for the broker and CSV output."""
-        self._endpoint_buffer = CoefficientEndpointBuffer(
-            redis_job_id, config_params.seed,
-            area, self.export_results_on_finish)
-
-        if self.export_results_on_finish:
-            self._export = CoefficientExportAndPlot(
-                area, self.export_path, self.export_subdir, self._endpoint_buffer)
-
-    def update_and_send_results(
-            self, current_state: dict, progress_info: SimulationProgressInfo,
-            area: "Area", simulation_status: str) -> None:
-        raise NotImplementedError
-
-    @classmethod
-    def _update_area_stats(cls, area: "Area", endpoint_buffer: "SimulationEndpointBuffer") -> None:
-        return
-
-    def update_csv_files(self, slot_no: int, current_time_slot: DateTime, area: "Area",
-                         scm_manager: SCMManager) -> None:
-        """Update the csv results on market cycle."""
-        if self.export_results_on_finish:
-            self._export.data_to_csv(area, current_time_slot, slot_no == 0, scm_manager)
-
-    def save_csv_results(self, area: "Area") -> None:
-        """Update the CSV results on finish, and write the CSV files."""
-        if self.export_results_on_finish:
-            log.info("Exporting simulation data.")
-            self._export.data_to_csv(area, False, None)
-            self._export.area_tree_summary_to_json(self._endpoint_buffer.area_result_dict)
-            self._export.export(power_flow=None)
-
-    def update_send_coefficient_results(
-            self, current_state: dict, progress_info: SimulationProgressInfo,
-            area: "CoefficientArea", simulation_status: str,
-            scm_manager: Optional["SCMManager"] = None) -> None:
-        # pylint: disable=too-many-arguments
-        """
-        Update the coefficient simulation results.
-        """
-        assert self._endpoint_buffer is not None
-
-        if self._should_send_results_to_broker:
-            self._endpoint_buffer.update_coefficient_stats(
-                area, simulation_status, progress_info, current_state,
-                False, scm_manager)
-            results = self._endpoint_buffer.prepare_results_for_publish()
-            if results is None:
-                return
-            self.kafka_connection.publish(results, current_state["simulation_id"])
-
-        elif (gsy_e.constants.RETAIN_PAST_MARKET_STRATEGIES_STATE or
-              self.export_results_on_finish):
-
-            self._endpoint_buffer.update_coefficient_stats(
-                area, current_state["sim_status"], progress_info, current_state,
-                True, scm_manager)
-            self._update_area_stats(area, self._endpoint_buffer)
-
-            if self.export_results_on_finish:
-                assert self._export is not None
-                if (progress_info.current_slot_time is not None
-                        and gsy_e.constants.RETAIN_PAST_MARKET_STRATEGIES_STATE):
-                    # for integration tests:
-                    self._export.raw_data_to_json(
-                        progress_info.current_slot_str,
-                        self._endpoint_buffer.flattened_area_core_stats_dict
-                    )
-
-                self._export.file_stats_endpoint(area, scm_manager)
-
-
-def simulation_results_manager_factory():
-    """Factory for results manager objects."""
-    if ConstSettings.MASettings.MARKET_TYPE == SpotMarketTypeEnum.COEFFICIENTS.value:
-        return CoefficientSimulationResultsManager
-    return SimulationResultsManager
-
-
-class SimulationExternalEvents:
-    """
-    Handle signals that affect the simulation state, that arrive from Redis. Consists of live
-    events and signals that change the simulation status.
-    """
-    def __init__(self, simulation_id: str, config: SimulationConfig,
-                 state_params: SimulationStatusManager, progress_info: SimulationProgressInfo,
-                 area: "Area") -> None:
-        # pylint: disable=too-many-arguments
-        self.live_events = LiveEvents(config)
-        self.redis_connection = RedisSimulationCommunication(
-            state_params, simulation_id, self.live_events, progress_info, area)
-
-    def update(self, area: "AreaBase") -> None:
-        """
-        Update the simulation according to any live events received. Triggered every market slot.
-        """
-        self.live_events.handle_all_events(area)
 
 
 class Simulation:
